+++ addon/services/image-service.js
@@ -1,5 +1,5 @@
-import Ember from 'ember';
+import Service from '@ember/service';
 import serviceMixin from '../mixins/ags-service-mixin';
 
-export default Ember.Service.extend(serviceMixin, {
+export default Service.extend(serviceMixin, {
 });