+++ ember-cli-build.js
@@ -1,9 +1,10 @@
+'use strict';
 /* jshint node:true */
 /* global require, module */
-var EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
+const EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
 
-module.exports = function (defaults) {
-  var app = new EmberAddon(defaults, {
+module.exports = function(defaults) {
+  let app = new EmberAddon(defaults, {
     // Add options here
   });