+++ tests/dummy/app/sites/site/route.js
@@ -1,7 +1,8 @@
-import Ember from 'ember';
+import { inject as service } from '@ember/service';
+import Route from '@ember/routing/route';
 import ENV from '../../config/environment';
-export default Ember.Route.extend({
-  featureService: Ember.inject.service('feature-service'),
+export default Route.extend({
+  featureService: service('feature-service'),
   model (params) {
     let url = ENV.APP.domainServiceUrl;
     let options = {